# api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Synckor, Koreader Sync Server
servers:
    - url: http://localhost:8050
      description: Development server
#    - url:
#      description: Production server

paths:
  /healthcheck:
    get:
      responses:
        '200':
          description: Server is up & running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /users/create:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPayload'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '402':
          description: Any unsuccessful response (User already exists, registration is disabled, etc)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /users/auth:
    get:
      parameters:
        - name: x-auth-user
          in: header
          required: true
          schema:
            type: string
            example: "username"
            description: Username of the user
        - name: x-auth-key
          in: header
          required: true
          schema:
            type: string
            example: "password"
            description: Password hash of the user
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          description: User authentication failed (inactive user, wrong password, etc)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /syncs/progress:
    put:
        parameters:
          - name: x-auth-user
            in: header
            required: true
            schema:
                type: string
                example: "username"
                description: Username of the user
          - name: x-auth-key
            in: header
            required: true
            schema:
                type: string
                example: "password"
                description: Password hash of the user
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPayload'
        responses:
            '200':
              description: Sync progress updated successfully
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/Response'
            '401':
              description: Sync progress update failed (inactive user, wrong password, etc)
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/Response'
components:
  schemas:
    UserPayload:
      type: object
      properties:
        username:
          type: string
          example: "anthony"
        password:
          type: string
          example: "tran"
    DocumentPayload:
      type: object
      properties:
        document:
          type: string
          example: "7e330d931b87f03326375cae"
        progress:
          type: string
          example: "0.5"
        percentage:
          type: number
          example: 0.5
        device:
          type: string
          example: "kindle-pw4"
        device_id:
          type: string
          example: "7e330d931b87f03326375cae"
    Response:
        type: object
        properties:
            userName:
              type: string
              example: "anthony"
            message:
              type: string
              example: "Server is up & running"